pyproject.toml
[project]
name = "jalaarogya"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.2",
    "pydantic>=2.11.9",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.1",
    "uvicorn>=0.35.0",
]
